Initial commit
withdraw_index returns 2 dataframes: exclusions for genotyped data
(based on the fam file) and imputed data (based on the sample file).
In addition samples with a value of 'redacted'/ 'dropout' in the
phenotype column of the fam file have been included in the
exclusions for both genotyped and imputed data.